[feat]: The Whole idea of the project change and products were added
- We are going to have more validations in the back no more methods doing
  specific things.
- The controllers methods in general validates if an user is admin.
- Migration, Model, Controller and Routes were added for products.

Resolves: #NA
See also: #NA

+++ app/controllers/origins_controller.rb
@@ -14,13 +14,21 @@ class OriginsController < ApplicationController
     save_and_render origin
   end
 
-  def update 
-    @origin.update_attributes origin_params 
-    save_and_render @origin
+  def update
+    if @origin.user_id == @current_user.id 
+      @origin.update_attributes origin_params 
+      save_and_render @origin
+    else
+      permissions_error
+    end
   end
 
   def destroy
-    render_ok @origin.destroy  
+    if @origin.user_id == @current_user.id 
+      render_ok @origin.destroy  
+    else
+      permissions_error
+    end
   end
 
   private
